rename core utils
